v 0.3. Bug fixed in save_integral_fluxes_to_file
Changed program so that outputs file containing integral fluxes and the  filename contains that date of the SEP event (rather than a date range). Fixed a bug so that the program won't crash even if no thresholds or only one flux threshold is crossed during the specified time period. Will not save an output file with integral fluxes if no thresholds are crossed (previously, was creating an output file no matter what).