v.4 Fixed imported module in loop prob. SHINE best
Importing my codes as modules in the wrapper and then running them in a loop caused problems. When the module sep was called in the loop, sometimes information from a previous run was retained or somehow corrupted the next run. Reloading the module each time in the loop fixed the problem. Uploading the fluence, integral flux time series, and sep values for all GOES-13 and GOES-15 measurements again, just in case.